+++ src/page/footer/style/index.ts
@@ -1,0 +1,12 @@
+import styled from "styled-components";
+
+
+export const StyledFooter = styled.footer`
+  position: fixed;
+  bottom: 0;
+  left: 0;
+  width: 100%;
+  background-color: ${({ theme }) => theme.paletteColor.lightGreen};
+  box-shadow: 0 -2px 10px rgba(0, 0, 0, 0.1);
+  z-index: 1000;
+`;
